Poprawiony obrót.

+++ GUI.h
@@ -13,8 +13,8 @@ struct ConfigContext{
      int p4;
      int p5;
      int p6;
-     int p7;
-     int p8;
+     float p7;
+     float p8;
 };
 
 void drawLeftPanel(ImGuiIO& io);
